move check_number_bounds to numeric, add some tests
